+++ src/Cli/CommandCollection.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Bauhaus\CliApplication;
+namespace Bauhaus\Cli;
 
 use ArrayIterator;
 use Bauhaus\CliEntrypoint;
